+++ packages/recipes/libfranka-fr3/run.sh
@@ -1,2 +1,0 @@
-cd ~/libfranka-0.16.1/build/examples
-./echo_robot_state 172.16.0.2